Rename azure runtime's package to com.microsoft.azure
