+++ azure-android-client-authentication/src/main/java/com/microsoft/azure/credentials/package-info.java
