+++ azure-client-runtime/src/main/java/com/microsoft/azure/CloudException.java
@@ -5,8 +5,9 @@
  *
  */
 
-package com.microsoft.rest;
+package com.microsoft.azure;
 
+import com.microsoft.rest.AutoRestException;
 import retrofit.Response;
 
 /**
